Conflicts:
	reformulation-core/src/main/java/it/unibz/krdb/obda/owlrefplatform/core/translator/SparqlAlgebraToDatalogTranslator.java

+++ obdalib-protege41/src/main/java/it/unibz/krdb/obda/protege4/views/QueryInterfaceView.java
@@ -21,12 +21,8 @@ package it.unibz.krdb.obda.protege4.views;
  */
 
 import it.unibz.krdb.obda.io.PrefixManager;
-import it.unibz.krdb.obda.model.OBDAException;
 import it.unibz.krdb.obda.model.impl.OBDAModelImpl;
 import it.unibz.krdb.obda.owlapi3.OWLResultSetWriter;
-import it.unibz.krdb.obda.owlrefplatform.core.queryevaluation.SPARQLQueryUtility;
-import it.unibz.krdb.obda.owlrefplatform.owlapi3.QuestOWL;
-import it.unibz.krdb.obda.owlrefplatform.owlapi3.QuestOWLConnection;
 import it.unibz.krdb.obda.owlrefplatform.owlapi3.QuestOWLResultSet;
 import it.unibz.krdb.obda.owlrefplatform.owlapi3.QuestOWLStatement;
 import it.unibz.krdb.obda.protege4.core.OBDAModelManager;
@@ -41,23 +37,6 @@ import it.unibz.krdb.obda.protege4.utils.DialogUtils;
 import it.unibz.krdb.obda.protege4.utils.OBDAProgessMonitor;
 import it.unibz.krdb.obda.protege4.utils.OBDAProgressListener;
 import it.unibz.krdb.obda.protege4.utils.TextMessageFrame;
-
-import java.awt.BorderLayout;
-import java.awt.Dimension;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.Writer;
-import java.util.List;
-import java.util.concurrent.CountDownLatch;
-
-import javax.swing.JFrame;
-import javax.swing.JOptionPane;
-import javax.swing.JPanel;
-import javax.swing.JSplitPane;
-import javax.swing.SwingUtilities;
-import javax.swing.table.DefaultTableModel;
-
 import org.protege.editor.core.ProtegeManager;
 import org.protege.editor.owl.OWLEditorKit;
 import org.protege.editor.owl.ui.view.AbstractOWLViewComponent;
@@ -65,9 +44,18 @@ import org.semanticweb.owlapi.model.OWLAxiom;
 import org.semanticweb.owlapi.model.OWLException;
 import org.semanticweb.owlapi.model.OWLOntologyChange;
 import org.semanticweb.owlapi.model.OWLOntologyChangeListener;
-import org.semanticweb.owlapi.reasoner.OWLReasoner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import javax.swing.*;
+import javax.swing.table.DefaultTableModel;
+import java.awt.*;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.Writer;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
 
 public class QueryInterfaceView extends AbstractOWLViewComponent implements SavedQueriesPanelListener, OBDAModelManagerListener {
 
@@ -155,7 +143,7 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 		setupListeners();
 		
 		// Setting up actions for all the buttons of this view.
-		resultTablePanel.setCountAllTuplesActionForUCQ(new OBDADataQueryAction<Integer>("Counting tuples...", QueryInterfaceView.this) {
+		resultTablePanel.setCountAllTuplesActionForUCQ(new OBDADataQueryAction<Long>("Counting tuples...", QueryInterfaceView.this) {
 			@Override
 			public OWLEditorKit getEditorKit(){
 				return getOWLEditorKit();
@@ -165,13 +153,13 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 				return -1;
 			}
 			@Override
-			public void handleResult(Integer result){
+			public void handleResult(Long result){
 				updateTablePanelStatus(result);	
 			}
 
 
 			@Override
-			public Integer executeQuery(QuestOWLStatement st, String query) throws OWLException {
+			public Long executeQuery(QuestOWLStatement st, String query) throws OWLException {
 				return st.getTupleCount(query);
 			}
 
@@ -366,8 +354,8 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 
 	
 	private class QueryStatusUpdater implements Runnable{
-		int result;
-		QueryStatusUpdater(int result){
+		long result;
+		QueryStatusUpdater(long result){
 			this.result = result;
 		}
 		
@@ -375,7 +363,7 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 			queryEditorPanel.updateStatus(result);
 		}
 	}
-	protected void updateTablePanelStatus(int result) {
+	protected void updateTablePanelStatus(Long result) {
 		if (result != -1) {
 			Runnable status_updater = new QueryStatusUpdater(result);
 			SwingUtilities.invokeLater(status_updater);
